logic for toggling icon

+++ src/services/bookmarks-services.ts
@@ -15,11 +15,14 @@ export const createBookmark = async (recipeInfo: RecipeType, user: string) => {
     }
     const newRecipeObj = await getRecipeByID(recipeInfo.id);
     const recipe = newRecipeObj._id;
-    const response = await api.post(`${BOOKMARKS_URL}`, {
-      user, recipe
-    });
-    const newBookmark = response.data;
-    return newBookmark;
+    if (recipe) {
+      const response = await api.post(`${BOOKMARKS_URL}`, {
+        user, recipe
+      });
+      const newBookmark = response.data;
+      return newBookmark;
+    }
+    return null;
   } catch (error) {
     console.log("error: ", error);
     return null;
@@ -30,7 +33,23 @@ export const getBookmarksByRecipeId = async (recipeId) => {
   try {
     const recipeObj = await getRecipeByID(recipeId);
     const recipeObjId = recipeObj._id;
-    const response = await api.get(`${BOOKMARKS_URL}/${recipeObjId}`);
+    if (recipeObjId) {
+      const response = await api.get(`${BOOKMARKS_URL}/${recipeObjId}`);
+      const bookmarks = response.data;
+      return bookmarks;
+    }
+    return null;
+  } catch (error) {
+    console.log("error: ", error);
+    return null;
+  }
+};
+
+export const getBookmarksByUserId = async (userId, objectsOrRecipes) => {
+  try {
+    const response = objectsOrRecipes ?
+        await api.get(`${BOOKMARKS_URL}/objects/user/${userId}/`) :
+        await api.get(`${BOOKMARKS_URL}/user/${userId}/`);
     const bookmarks = response.data;
     return bookmarks;
   } catch (error) {
@@ -39,11 +58,16 @@ export const getBookmarksByRecipeId = async (recipeId) => {
   }
 };
 
-export const getBookmarksByUserId = async (userId) => {
+export const getBookmarksByBothIds = async (userId, recipeId) => {
   try {
-    const response = await api.get(`${BOOKMARKS_URL}/user/${userId}`);
-    const bookmarks = response.data;
-    return bookmarks;
+    const recipeObj = await getRecipeByID(recipeId);
+    const recipeObjId = recipeObj._id;
+    if (recipeObjId) {
+      const response = await api.get(`${BOOKMARKS_URL}/${recipeObjId}/${userId}/`);
+      const bookmarks = response.data;
+      return bookmarks;
+    }
+    return null;
   } catch (error) {
     console.log("error: ", error);
     return null;
@@ -54,9 +78,11 @@ export const unbookmark = async (recipeInfo: RecipeType, user: string) => {
   try {
     const recipeObj = await getRecipeByID(recipeInfo.id);
     const recipe = recipeObj._id;
-    const bookmark = { user: user, recipe: recipe };
-    const deletion = await api.delete(`${BOOKMARKS_URL}`, { data: { user: user, recipe: recipe } });
-
+    if (recipe) {
+      const response = await api.delete(`${BOOKMARKS_URL}`, { data: { user: user, recipe: recipe } });
+      return response.data;
+    }
+    return null;
   } catch (error) {
     console.log("error: ", error);
     return null;
